Add support for EPSV.
Extended passive mode is similar to passive, but it works over IPv6.
The server only response with a port, and then will listen v4/v6
depending on what the control connection is. There doesn't seem to
be a FEAT for this functionality, so the goftp will always try
EPSV first, then fall back to PASV.