Add Store() method for uploading files.
It also supports resuming interrupted uploads if you pass it an io.Reader
that is also an io.Seeker (since it has to be able to rewind the reader
based on how much the server actually received before it was interrupted).
It won't resume uploads if your client is connected to different servers
because it wouldn't work, and it doesn't really make sense to connect
to multiple servers to perform uploads.

I also added a Close() method you can use to close a client and all its
connections. The tests use this to simulate server side issues instead
of restarting the server since that didn't work very well.

I also changed the pure-ftpd build script to activate the puredb feature
which allows creating arbitrary user accounts for the ftp server. I
added a "goftp" user so I could test resuming uploads (pure-ftpd
doesn't let anonymous users write to existing files).