Add ReadDir() function.
This uses the MLSD feature to get more details about the files in a directory.
It mirrors the ioutil.ReadDir function. I named the other API functions based
on the FTP command names, but this time I made an executive decision not to
name this function "MachineListDir".

Also start splitting things up into separate files.