Improve ReadDir/Stat support.
If the server doesn't support MLSD/MLST, try to parse the output of
LIST. Only "ls -l" output is supported. The mtime in the output doesn't
include the time zone, so I added a config field to specify the server's
time zone if your mtimes are that important to you (but they won't be
very precise anyway).